+++ nd2k/types.py
@@ -5,6 +5,7 @@ from enum import Enum
 from .operation import OperationType, Operation
 from .nontrade import NonTrade
 from .trade import Trade
+from .swap import Swap
 
 
 CSV = list[list[str]]
@@ -23,33 +24,6 @@ class KoinlyTag(Enum):
 	SWAP            = "swap"
 	EXCHANGE        = "exchange"
 	EXCHANGE_FEE    = "fee"
-
-
-@dataclass
-class Swap:
-	asset_a: Operation
-	asset_b: Operation
-
-	@property
-	def date(self) -> datetime:
-		return self.asset_a.date
-
-	@property
-	def summary(self) -> str:
-		return "".join([
-			self.asset_a.summary,
-			self.asset_a.symbol,
-			"/",
-			self.asset_b.symbol
-		])
-
-
-class PartialSwap:
-	def __init__(self, asset_a: Operation):
-		self.asset_a = asset_a
-
-	def complete(self, asset_b: Operation) -> Swap:
-		return Swap(asset_a=self.asset_a, asset_b=asset_b)
 
 
 @dataclass
